requirements.txt
MySQL-python==1.2.5
Twisted==13.2.0
zope.interface==4.0.5
PyYAML==3.11
mock=1.0.1
